requirements.txt
streamlit
openai
python-dotenv
langchain-openai
langchain-core
langchain-community
langgraph
pydantic
duckduckgo-search
ddgs